pyproject.toml
[build-system]
requires = ["setuptools>=77.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "scyjava"
version = "1.12.1.dev0"
description = "Supercharged Java access from Python"
license = "Unlicense"
authors = [{name = "SciJava developers", email = "ctrueden@wisc.edu"}]
readme = "README.md"
keywords = ["java", "maven", "cross-language"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: Unix",
    "Operating System :: MacOS",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries :: Java Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]

# NB: Keep this in sync with environment.yml AND dev-environment.yml!
requires-python = ">=3.9"
dependencies = ["jpype1 >= 1.3.0", "jgo", "cjdk", "stubgenj"]

[project.optional-dependencies]
# NB: Keep this in sync with dev-environment.yml!
dev = [
    "assertpy",
    "build",
    "jep",
    "pytest",
    "pytest-cov",
    "numpy",
    "pandas",
    "ruff",
    "toml",
    "validate-pyproject[all]",
]

[project.scripts]
scyjava-stubgen = "scyjava._stubs._cli:main"

[project.entry-points.hatch]
scyjava = "scyjava._stubs._hatchling_plugin"
[project.entry-points."distutils.commands"]
build_py = "scyjava_stubgen.build:build_py"

[project.urls]
homepage = "https://github.com/scijava/scyjava"
documentation = "https://github.com/scijava/scyjava/blob/main/README.md"
source = "https://github.com/scijava/scyjava"
download = "https://pypi.org/project/scyjava/"
tracker = "https://github.com/scijava/scyjava/issues"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = false

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false

# ruff configuration
[tool.ruff]
line-length = 88
src = ["src", "tests"]
include = ["pyproject.toml", "src/**/*.py", "tests/**/*.py"]
extend-exclude = ["bin", "build", "dist"]

[tool.ruff.lint]
extend-ignore = ["E203"]

[tool.ruff.lint.per-file-ignores]
# Ignore `E402` (import violations) in all `__init__.py` files, and in `path/to/file.py`.
"__init__.py" = ["E402", "F401"]
